refactor: cursor 생성 DTO import 경로 수정
